Add files via upload
My current interest revolves around developing quantum optimization algorithms and error correction for NISQ computers. Through my research I came across a novel yet simple problem: If Alice was to send Bob a qubit in a particular state, lets say state |0> for this example, how would Bob verify that he received the correct state (the state that Alice sent)?
The simplest solution to this problem involves sending two additional qubits with the corresponding qubit. Upon arrival, a simple logic circuit consisting of controlled-NOT (CX) gates is applied. The measurement of each out come is examined and allows to make a fair judgement on the original qubit. Please see the code for more.